pyproject.toml
[project]
name = "telegram-bot-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot",
]